+++ Shared/Game.Shared.External/Providers/Ai/LlamaCpp/Internal/ChatResult.cs
